Use size_t for map sizes

+++ include/map.h
@@ -1,13 +1,15 @@
 #pragma once
 
+#include <stdlib.h>
+
 struct mushroom_map_item {
 	char *key;
 	char *value;
 };
 
 struct mushroom_map {
-	int size;
-	int count;
+	size_t size;
+	size_t count;
 	struct mushroom_map_item **items;
 };
